wallet: static cast `nHashType` function argument from `int` to `uint8_t`
This is done to fix failures in the fuzz tests that occur because of implicit
conversion of `nHashType` from `int` to `uint8_t` when ultimately passed down
 to the `SignatureHashSchnorr` function. Narrowing the type for this variable
at this stage fixes the fuzz tests failures.

+++ src/script/sign.cpp
@@ -81,7 +81,7 @@ std::optional<uint256> MutableTransactionSignatureCreator::ComputeSchnorrSignatu
         execdata.m_tapleaf_hash = *leaf_hash;
     }
     uint256 hash;
-    if (!SignatureHashSchnorr(hash, execdata, m_txto, nIn, nHashType, sigversion, *m_txdata, MissingDataBehavior::FAIL)) return std::nullopt;
+    if (!SignatureHashSchnorr(hash, execdata, m_txto, nIn, static_cast<uint8_t>(nHashType), sigversion, *m_txdata, MissingDataBehavior::FAIL)) return std::nullopt;
     return hash;
 }
